Cleanup test command
